off実行後にステータス表示オプション-v追加
